refactor(release): share the strip list and pack helper for third consumers
PR #314's check-published-diff reproduces the publish workflow's Pack
step with its own copies of the manifest strip and the pnpmPack helper,
under a "keep in lockstep" comment — the manual-sync hazard this
refactor exists to kill. Export STRIPPED_MANIFEST_FIELDS from
check-pack.core.ts (the exact `npm pkg delete` list, now driving the
packed-manifest violations and locked by a test) and extract the
corepack-fallback pnpmPack into scripts/pack.ts so both check-pack.ts
and check-published-diff.ts can import one copy. check-published-diff
switches over in a follow-up once both PRs land.
